Refactor `functions` to `feed_processing`
