--- functions/rss_files/nonlinear-library-AF_sample.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<rss xmlns:dc="http://purl.org/dc/elements/1.1/"
-     xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
-    <channel>
-        <title>
-            <![CDATA[AI Alignment Forum]]>
-        </title>
-        <description>
-            <![CDATA[A community blog devoted to technical AI alignment research]]>
-        </description>
-        <link>https://www.alignmentforum.org</link>
-        <image>
-            <url>https://res.cloudinary.com/lesswrong-2-0/image/upload/v1497915096/favicon_lncumn.ico</url>
-            <title>AI Alignment Forum</title>
-            <link>https://www.alignmentforum.org</link>
-        </image>
-        <generator>RSS for Node</generator>
-        <lastBuildDate>Tue, 11 Apr 2023 10:17:06 GMT</lastBuildDate>
-        <atom:link href="https://www.alignmentforum.org/feed.xml?view=community-rss&amp;karmaThreshold=0" rel="self"
-                   type="application/rss+xml"/>
-        <item>
-            <title>
-                <![CDATA[AF - Expanding the domain of discourse reveals structure already there but hidden by Tsvi Benson-Tilsen]]>
-            </title>
-            <description>
-                <![CDATA[ Welcome to The Nonlinear Library, where we use Text-to-Speech software to convert the best writing from the Rationalist and EA communities into audio.]]>
-            </description>
-            <link>
-                https://www.alignmentforum.org/posts/smDeWfgeYDg9eGq5G/environments-for-measuring-deception-resource-acquisition
-            </link>
-            <guid isPermaLink="false">smDeWfgeYDg9eGq5G_NL_AF</guid>
-            <dc:creator>
-                <![CDATA[Dan H]]>
-            </dc:creator>
-            <pubDate>Fri, 07 Apr 2023 18:40:22 GMT</pubDate>
-        </item>
-    </channel>
-</rss>
