--- functions/rss_files/nonlinear-library-EA-empty.xml
@@ -1,18 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<rss
-        xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
-    <channel>
-        <title><![CDATA[Effective Altruism Forum]]></title>
-        <description><![CDATA[Effective Altruism research, discussion and community updates]]></description>
-        <link>https://forum.effectivealtruism.org</link>
-        <image>
-            <url>https://res.cloudinary.com/lesswrong-2-0/image/upload/v1497915096/favicon_lncumn.ico</url>
-            <title>Effective Altruism Forum</title>
-            <link>https://forum.effectivealtruism.org</link>
-        </image>
-        <generator>RSS for Node</generator>
-        <lastBuildDate>Sat, 22 Apr 2023 22:00:05 GMT</lastBuildDate>
-        <atom:link href="https://forum.effectivealtruism.org/feed.xml?view=community-rss&amp;karmaThreshold=25"
-                   rel="self" type="application/rss+xml"/>
-    </channel>
-</rss>
