--- functions/rss_files/nonlinear-library-LW-empty.xml
@@ -1,6 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<rss xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:content="http://purl.org/rss/1.0/modules/content/" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
-  <channel><title><![CDATA[LessWrong]]></title>        <description><![CDATA[A community blog devoted to refining the art of rationality]]></description><link>https://www.lesswrong.com</link>        <image><url>https://res.cloudinary.com/lesswrong-2-0/image/upload/v1497915096/favicon_lncumn.ico</url>        <title>LessWrong</title><link>https://www.lesswrong.com</link></image><generator>RSS for Node</generator>        <lastBuildDate>Sat, 22 Apr 2023 22:00:06 GMT</lastBuildDate><atom:link href="https://www.lesswrong.com/feed.xml?view=community-rss&amp;karmaThreshold=30" rel="self" type="application/rss+xml"/><item><title><![CDATA[LW - ask me about my battery by bhauth]]></title><description>        <![CDATA[ Welcome to The Nonlinear Library, where we use Text-to-Speech software to convert the best writing from the Rationalist and EA communities into audio.
-    This is: ask me about my battery, published by bhauth on April 21, 2023 on LessWrong. <br /><br /><p>Some years ago, I designed a battery chemistry. Then, I got a semi-related chemical industry job. Then, in 2020, I did some testing at a university lab. Here's the basic design:</p><figure class="image"><img src="https://res.cloudinary.com/lesswrong-2-0/image/upload/f_auto,q_auto/v1/mirroredImages/Xn8EjGzndLbAsadW2/q3gz4t2qjmgdkucxtwho" /></figure><p>As far as I was able to test it, the battery design worked, but the project died. I haven't been able to do much with that design since.</p><p>The current name for this battery chemistry is "SMAC" for "sodium metal aqueous cathode". If you want, you can ask about:</p><ul><li>technical details</li><li>experimental results</li><li>why I didn't do X</li><li>the response from X types of people</li><li>other battery chemistries</li></ul><p>I could certainly provide more details here, but I'm curious what people are most interested in.</p> <p>Thanks for listening. To help us out with The Nonlinear Library or to learn more, please visit nonlinear.org. </p><br /><br /><a href="https://www.lesswrong.com/posts/Xn8EjGzndLbAsadW2/ask-me-about-my-battery#comments">Discuss</a>]]></description>
-        <link>https://www.lesswrong.com/posts/Xn8EjGzndLbAsadW2/ask-me-about-my-battery</link><guid isPermaLink="false">Xn8EjGzndLbAsadW2_NL_LW</guid><dc:creator><![CDATA[bhauth]]></dc:creator><pubDate>Sat, 22 Apr 2023 17:31:03 GMT</pubDate></channel>
-</rss>
