+++ manual_tests/run_beyondwords_feed_all.py
@@ -1,5 +1,5 @@
-from functions.configs import beyondwords_af_config, beyondwords_ea_config, beyondwords_lw_config
-from functions.feed_updaters import update_beyondwords_input_feed
+from feed_processing.configs import beyondwords_af_config, beyondwords_ea_config, beyondwords_lw_config
+from feed_processing.feed_updaters import update_beyondwords_input_feed
 
 if __name__ == '__main__':
     update_beyondwords_input_feed(beyondwords_af_config(), False)
